refactor: add util package and change directory
